Home

EffectSizes.jl

EffectSizes.EffectSizes

EffectSizes.AbstractEffectSize

EffectSizes.CohenD

EffectSizes.ConfidenceInterval

EffectSizes.EffectSize

EffectSizes.GlassΔ

EffectSizes.HedgeG

EffectSizes.lower

EffectSizes.upper

Statistics.quantile

StatsBase.confint

EffectSizes.EffectSizes — Module.

EffectSizes

A Julia package for effect size measures.

source

EffectSizes.CohenD — Type.

CohenD(xs, ys[, bootstrap]; [quantile=0.95])

Calculate Cohen's $d$ effect size index between two vectors xs and ys.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ d = \frac{m_A - m_B}{s}\]

where $m$ is the mean and $s$ is the pooled standard deviation:

\[ s = \sqrt{\frac{(n_A - 1) s_A^2 + (n_B - 1) s_B^2}{n_A + n_B - 2}}\]

If $m_A$ > $m_B$, $d$ will be positive and if $m_A$ < $m_B$, $d$ will be negative.

Note

HedgeG outperforms CohenD when sample sizes are < 20.

Examples

xs = randn(100000)
ys = randn(100000) .+ 0.01

using EffectSizes
CohenD(xs, ys)

using HypothesisTests
EqualVarianceTTest(xs, ys)

source

EffectSizes.ConfidenceInterval — Type.

ConfidenceInterval(lower, upper, quantile)

A type representing the lower lower and upper upper bounds of an effect size confidence interval at the quantile quantile.

ConfidenceInterval(xs, ys, d; quantile)
ConfidenceInterval(xs, ys; quantile, bootstrap)

Calculates the effect size confidence interval between two vectors xs and ys at the quantile quantile.

If d is provided, the confidence interval is calculated using the standard normal distribution. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

source

EffectSizes.EffectSize — Type.

const EffectSize = CohenD

See CohenD.

source

EffectSizes.GlassΔ — Type.

GlassΔ(treatment, control[, bootstrap]; [quantile=0.95])

Calculate Glass's $Δ$ effect size index between two vectors treatment and control.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ Δ = \frac{m_T - m_C}{s_C}\]

where $m$ is the mean, $s$ is the standard deviation, $T$ is the treatment group and $C$ is the control group.

If $m_T$ > $m_C$, $Δ$ will be positive and if $m_T$ < $m_C$, $Δ$ will be negative.

Note

GlassΔ should be used when the standard deviations between the two groups are very different.

source

EffectSizes.HedgeG — Type.

HedgeG(xs, ys[, bootstrap]; [quantile=0.95])

Calculate Hedge's $g$ effect size index between two vectors xs and ys.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ g = \frac{m_A - m_B}{s}\]

where $m$ is the mean and $s$ is the pooled standard deviation:

\[ s = \sqrt{\frac{(n_A - 1) s_A^2 + (n_B - 1) s_B^2}{n_A + n_B}}\]

If $m_A$ > $m_B$, $g$ will be positive and if $m_A$ < $m_B$, $g$ will be negative.

Note

HedgeG outperforms CohenD when sample sizes are < 20.

source

EffectSizes.lower — Method.

lower(ci)

Returns the lower bound of a confidence interval.

source

EffectSizes.upper — Method.

upper(ci)

Returns the upper bound of a confidence interval.

source

Statistics.quantile — Method.

quantile(ci)

Returns the quantile of a confidence interval.

source

StatsBase.confint — Method.

confint(ci)

Returns the lower and upper bounds of a confidence interval.

source

EffectSizes.AbstractEffectSize — Type.

AbstractEffectSize

An abstract type to represent an effect size.

Effect	Effect size

Small	0.2

Medium	0.5

Large	0.8

Subtypes implement:

Method	Description

effectsize	returns the effect size index

confint	returns the confidence interval

quantile	returns the confidence interval quantile

source