Home

EffectSizes.jl

EffectSizes.jl is a Julia package for effect size measures. Confidence intervals are assigned to effect sizes using either the normal distribution or by bootstrap resampling. The package implements types for the following measures:

Measure	Type

Cohen's d	CohenD

Hedge's g	HedgeG

Glass's Δ	GlassΔ

Installation

] add https://github.com/harryscholes/EffectSizes.jl

Examples

julia> using Random, EffectSizes; Random.seed!(1);

julia> xs = randn(10^3);

julia> ys = randn(10^3) .+ 0.5;

julia> es = CohenD(xs, ys) # normal CI
-0.507, 0.95CI (-0.924, -0.089)

julia> effectsize(es)
-0.506674937960848

julia> quantile(es)
0.95

julia> CohenD(xs, ys, quantile=0.99)
-0.507, 0.99CI (-1.056, 0.042)

julia> CohenD(xs, ys, 10^4) # bootstrap CI
-0.507, 0.95CI (-0.594, -0.417)

julia> ci = confint(es)
0.95CI (-0.924, -0.089)

julia> confint(ci)
(-0.9244427501651218, -0.08890712575657417)

julia> lower(ci)
-0.9244427501651218

julia> upper(ci)
-0.08890712575657417

julia> quantile(ci)
0.95

Index

EffectSizes.AbstractEffectSize

EffectSizes.BootstrapConfidenceInterval

EffectSizes.CohenD

EffectSizes.ConfidenceInterval

EffectSizes.EffectSize

EffectSizes.GlassΔ

EffectSizes.HedgeG

EffectSizes.effectsize

EffectSizes.lower

EffectSizes.upper

Statistics.quantile

StatsBase.confint

API

EffectSizes.AbstractEffectSize — Type.

AbstractEffectSize

An abstract type to represent an effect size.

Effect	Effect size

Small	0.2

Medium	0.5

Large	0.8

Subtypes implement:

Method	Description

effectsize	returns the effect size index

confint	returns the confidence interval

quantile	returns the confidence interval quantile

source

EffectSizes.CohenD — Type.

CohenD(xs, ys[, bootstrap]; [quantile=0.95])

Calculate Cohen's $d$ effect size index between two vectors xs and ys.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ d = \frac{m_A - m_B}{s}\]

where $m$ is the mean and $s$ is the pooled standard deviation:

\[ s = \sqrt{\frac{(n_A - 1) s_A^2 + (n_B - 1) s_B^2}{n_A + n_B - 2}}\]

If $m_A$ > $m_B$, $d$ will be positive and if $m_A$ < $m_B$, $d$ will be negative.

Note

HedgeG outperforms CohenD when sample sizes are < 20.

Examples

xs = randn(100000)
ys = randn(100000) .+ 0.01

using EffectSizes
CohenD(xs, ys)

using HypothesisTests
EqualVarianceTTest(xs, ys)

source

EffectSizes.EffectSize — Type.

const EffectSize = CohenD

See CohenD.

source

EffectSizes.HedgeG — Type.

HedgeG(xs, ys[, bootstrap]; [quantile=0.95])

Calculate Hedge's $g$ effect size index between two vectors xs and ys.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ g = \frac{m_A - m_B}{s}\]

where $m$ is the mean and $s$ is the pooled standard deviation:

\[ s = \sqrt{\frac{(n_A - 1) s_A^2 + (n_B - 1) s_B^2}{n_A + n_B}}\]

If $m_A$ > $m_B$, $g$ will be positive and if $m_A$ < $m_B$, $g$ will be negative.

Note

HedgeG outperforms CohenD when sample sizes are < 20.

source

EffectSizes.GlassΔ — Type.

GlassΔ(treatment, control[, bootstrap]; [quantile=0.95])

Calculate Glass's $Δ$ effect size index between two vectors treatment and control.

A confidence interval for the effect size is calculated at the quantile quantile. If bootstrap is provided, the confidence interval is calculated by resampling from xs and ys bootstrap times.

\[ Δ = \frac{m_T - m_C}{s_C}\]

where $m$ is the mean, $s$ is the standard deviation, $T$ is the treatment group and $C$ is the control group.

If $m_T$ > $m_C$, $Δ$ will be positive and if $m_T$ < $m_C$, $Δ$ will be negative.

Note

GlassΔ should be used when the standard deviations between the two groups are very different.

source

EffectSizes.ConfidenceInterval — Type.

ConfidenceInterval(lower, upper, quantile)

A type representing the lower lower and upper upper bounds of an effect size confidence interval at the quantile quantile.

ConfidenceInterval(xs, ys, es; quantile)

Calculates a confidence interval for the effect size es between two vectors xs and ys at the quantile quantile.

source

EffectSizes.BootstrapConfidenceInterval — Type.

BootstrapConfidenceInterval(lower, upper, quantile, bootstrap)

A type representing the lower lower and upper upper bounds of an effect size confidence interval at the quantile quantile with bootstrap boostrap resamples.

BootstrapConfidenceInterval(xs, ys, f, bootstrap; quantile)

Calculates the effect size confidence interval between two vectors xs and ys at the quantile quantile by applying f to bootstrap bootstrap resamples of xs and ys.

source

EffectSizes.effectsize — Function.

effectsize(es::AbstractEffectSize)

Returns the effect size index.

source

StatsBase.confint — Function.

confint(es::AbstractEffectSize)

Returns the confidence interval of an effect size as a ConfidenceInterval object.

confint(ci::ConfidenceInterval)

Returns the lower and upper bounds of a confidence interval.

source

EffectSizes.lower — Function.

lower(ci::AbstractConfidenceInterval)

Returns the lower bound of a confidence interval.

source

EffectSizes.upper — Function.

upper(ci::AbstractConfidenceInterval)

Returns the upper bound of a confidence interval.

source

Statistics.quantile — Function.

quantile(es::AbstractEffectSize)
quantile(ci::ConfidenceInterval)

Returns the quantile of a confidence interval.

source